ai_api/requirements.txt
fastapi
uvicorn[standard]
openai
Pillow
qrcode
requests